scripts/exploration/requirements.txt
plotly.express
plotly
matplotlib
scipy
scikit-learn
pandas
azure.storage.blob
pyarrow